# kvs-watch: only fetch new data for appends