… doc/man3/flux_kvs_lookup.rst