Problem: The FLUX_KVS_WATCH_APPEND flag is implemented inefficiently in kvs-watch.  Everytime new data is appended, the entire contents of the watched key are retrieved and only new data calculated via an offset is sent to the watcher.  This significantly slows down performance of large appended data (such as stdio).

Solution: Instead of retrieving the entire contents of the watched key, fetch the tree object for the key from the KVS.  With access to the tree object's blobref array, we need only access the new appended blobrefs from the content store.  This significantly cuts down on the amount of data transfered during a kvs-watch.

Fixes #6414

---

side notes:

- so protocol wise
  - before it was
    - first RPC get all KVS data, return it all, note how long the data is
    - later RPCs, get all KVS data, only send new data past offset of prior request
  - now it is
    - RPCs now get KVS treeobj
      - keep track of how many references are in the array
      - get each blobref data from the content store
      - return that data to the user

So RPC wise the pros are:

- no longer iterating / building larger RPCs than is necessary

the cons are

- adding a level of RPC indirection (getting treeobj)
- possibly sending more RPCs back to caller (first RPC does not "combine" data into a single return)

For large data, this is a significant performance improvement for use cases like this:

`time flux run flux lptest 2000000 80`

wallclock time is generally being cut down from 58 seconds to 24 seconds.  It's a big win.  (FWIW, this was on corona.  On my laptop it was 15s before 11.7s afterwards, not as much as a win there, but still that's 22%)

For everything else (`flux job attach JOBID` for a finished job, job throughput), it appears that performance is a wash.  That the wins balance out the losses.  This is of course with limited testing in single node instances.

Edit: Also as an aside, there is more potential for support of #6292 with this implementation