@chu11 2024-11-18
re-pushed fixing up a lot of the minor things listed above.  I went with the general rule of making more descriptive errors when an error would be half-likely, but kept "function name: strerror" log messages for things that should extremely unlikely.