@chu11 2024-11-21
re-pushed with fixes per comments above.  Also added a few new cleanups.  I could probably split those out at this point in time.