@chu11 2024-12-04
re-pushed with fixes based on comments above

there are two new commits, one to deal with parameters on different lines and one to add a `treeobj_type_name()` convenience function.  So probably warrants another look @garlick